Did more refactoring to redux. Also made some cosmetic changes to the module.
